PR: session package renamed to devportalservice; GetSession() renamed to SessionData()
